#45 feat: Lending페이지 > Management 페이지 이름 변환
- 나머지 페이지 시멘틱 작업
