--- src/Pages/Lending.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-
-const Root = styled.div``;
-
-export default function LendingPage() {
-  return <Root>랜딩 페이지</Root>;
-}
