+++ src/Pages/MakingProject.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function MakingProjectPage() {
   return <Root>프로젝트 생성 페이지</Root>;
