+++ src/Pages/MakingTodo.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function MakingTodoPage() {
   return <Root>할일 생성 페이지</Root>;
