+++ src/Pages/Mypage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function MyPage() {
   return <Root>마이 페이지</Root>;
