+++ src/Pages/Profile.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function ProfilePage() {
   return <Root>프로필 페이지</Root>;
