+++ src/Pages/ProjectCategory.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function ProjectCategoryPage() {
   return <Root>프로젝트 카테고리 페이지</Root>;
