+++ src/Pages/ReplacePw.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function ReplacePwPage() {
   return <Root>비밀번호 재설정 페이지</Root>;
