+++ src/Pages/SetAlert.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function SetAlertPage() {
   return <Root>알림 설정 페이지</Root>;
