+++ src/Pages/Signup.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function SignupPage() {
   return <Root>회원가입 페이지</Root>;
