+++ src/Pages/TOS.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function TOSPage() {
   return <Root>이용약관 페이지</Root>;
