+++ src/Pages/Weekly.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-const Root = styled.div``;
+const Root = styled.main``;
 
 export default function WeeklyPage() {
   return <Root>주간 평가 페이지</Root>;
