+++ src/Router/Router.jsx
@@ -3,6 +3,7 @@ import { BrowserRouter, Route, Switch } from "react-router-dom";
 import styled from "styled-components";
 import LoginPage from "../Pages/Login";
 import MainPage from "../Pages/Main";
+import ManagementPage from "../Pages/Management ";
 import Test from "../Pages/Test";
 
 const ProjectWrapper = styled.div`
@@ -18,6 +19,7 @@ const Router = () => {
       <ProjectWrapper>
         <Switch>
           <Route exact path="/" component={MainPage} />
+          <Route exact path="/management" component={ManagementPage} />
           <Route path="/login" component={LoginPage} />
           <Route path="/test" component={Test} />
         </Switch>
